backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Storage
boto3==1.34.42
minio==7.2.4

# AI Libraries
torch==2.2.0
torchvision==0.17.0
diffusers==0.26.3
transformers==4.37.2
accelerate==0.26.1

# Video Processing
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
pillow==10.2.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
